Add JWT token validation to SubmissionController

+++ src/main/java/dev/hiruna/edutrack/controller/SubmissionController.java
@@ -2,7 +2,10 @@ package dev.hiruna.edutrack.controller;
 
 import dev.hiruna.edutrack.dto.SubmissionDTO;
 import dev.hiruna.edutrack.service.SubmissionService;
+import dev.hiruna.edutrack.util.JWTAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,28 +17,50 @@ public class SubmissionController {
     @Autowired
     private SubmissionService submissionService;
 
+    @Autowired
+    private JWTAuthenticator jwtAuthenticator;
+
     @GetMapping
-    public List<SubmissionDTO> getAllSubmissions() {
-        return submissionService.getAllSubmissions();
+    public ResponseEntity<List<SubmissionDTO>> getAllSubmissions(@RequestHeader("Authorization") String authHeader) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            return new ResponseEntity<>(submissionService.getAllSubmissions(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/{id}")
-    public SubmissionDTO getSubmissionById(@PathVariable Integer id) {
-        return submissionService.getSubmissionById(id);
+    public ResponseEntity<SubmissionDTO> getSubmissionById(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            SubmissionDTO submissionDTO = submissionService.getSubmissionById(id);
+            return new ResponseEntity<>(submissionDTO, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PostMapping
-    public SubmissionDTO createSubmission(@RequestBody SubmissionDTO submissionDTO) {
-        return submissionService.createSubmission(submissionDTO);
+    public ResponseEntity<SubmissionDTO> createSubmission(@RequestHeader("Authorization") String authHeader, @RequestBody SubmissionDTO submissionDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            SubmissionDTO createdSubmission = submissionService.createSubmission(submissionDTO);
+            return new ResponseEntity<>(createdSubmission, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/{id}")
-    public SubmissionDTO updateSubmission(@PathVariable Integer id, @RequestBody SubmissionDTO submissionDTO) {
-        return submissionService.updateSubmission(id, submissionDTO);
+    public ResponseEntity<SubmissionDTO> updateSubmission(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id, @RequestBody SubmissionDTO submissionDTO) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            SubmissionDTO updatedSubmission = submissionService.updateSubmission(id, submissionDTO);
+            return new ResponseEntity<>(updatedSubmission, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSubmission(@PathVariable Integer id) {
-        submissionService.deleteSubmission(id);
+    public ResponseEntity<Void> deleteSubmission(@RequestHeader("Authorization") String authHeader, @PathVariable Integer id) {
+        if (jwtAuthenticator.validateJwtToken(authHeader)) {
+            submissionService.deleteSubmission(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
